backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dotenv==1.0.1
setuptools==69.5.1
sqlparse==0.5.0
wheel==0.43.0
